backend/requirements.txt
# Python dependencies for the FlowWatch AI backend.
# Pin versions before production deployment.

# ── API ────────────────────────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# ── Data validation ────────────────────────────────────────────────────────────
pydantic>=2.6.0

# ── LLM / AI ──────────────────────────────────────────────────────────────────
anthropic>=0.25.0

# ── ML / Numeric ──────────────────────────────────────────────────────────────
# CPU-only install: pip install torch torchvision --index-url https://download.pytorch.org/whl/cpu
torch>=2.2.0
torchvision>=0.17.0
scikit-learn>=1.4.0
numpy>=1.26.0
pandas>=2.2.0
joblib>=1.3.0          # Model serialisation (also a scikit-learn dependency)

# ── AWS ────────────────────────────────────────────────────────────────────────
boto3>=1.34.0          # AWS SDK: Kinesis, CloudWatch, S3

# ── Database ──────────────────────────────────────────────────────────────────
asyncpg>=0.29.0        # Async PostgreSQL driver (TimescaleDB)

# ── Cache ─────────────────────────────────────────────────────────────────────
redis[asyncio]>=5.0.0

# ── Logging ───────────────────────────────────────────────────────────────────
loguru>=0.7.0

# ── Configuration ─────────────────────────────────────────────────────────────
python-dotenv>=1.0.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0          # Async test client for FastAPI
